+++ src/main/java/dk/kb/license/api/v1/impl/DsRightsApiServiceImpl.java
@@ -57,10 +57,12 @@ public class DsRightsApiServiceImpl extends ImplBase implements DsRightsApi {
     }
 
     @Override
-    public void deleteRestrictedId(String id, String idType, String platform, Boolean touchRecord) {
-        log.debug("Deleted restricted id:{} idType:{} platform:{}", id, idType, platform);
+    public RecordsCountDto deleteRestrictedId(String internalId, Boolean touchRecord) {
+        log.debug("Deleted restricted id from internalId: '{}'.", internalId);
         try {
-            RightsModuleFacade.deleteRestrictedId(id, idType, platform, getCurrentUserID(),touchRecord);
+            RecordsCountDto count = new RecordsCountDto();
+            count.setCount(RightsModuleFacade.deleteRestrictedId(internalId, getCurrentUserID(), touchRecord));
+            return count;
         } catch (Exception e) {
             throw handleException(e);
         }
@@ -99,9 +101,11 @@ public class DsRightsApiServiceImpl extends ImplBase implements DsRightsApi {
     }
 
     @Override
-    public void deleteRestrictedIds(List<RestrictedIdInputDto> restrictedIds, Boolean touchRecord) {
+    public RecordsCountDto deleteRestrictedIds(List<RestrictedIdInputDto> restrictedIds, Boolean touchRecord) {
         try {
-            RightsModuleFacade.deleteRestrictedIds(restrictedIds, getCurrentUserID(),touchRecord);
+            RecordsCountDto count = new RecordsCountDto();
+            count.setCount(RightsModuleFacade.deleteRestrictedIds(restrictedIds, getCurrentUserID(),touchRecord));
+            return count;
         } catch (Exception e) {
             throw handleException(e);
         }