+++ src/main/java/dk/kb/license/facade/RightsModuleFacade.java
@@ -95,28 +95,33 @@ public class RightsModuleFacade {
     }
 
     /**
-     * Delete the resticted id
+     * Delete a restricted ID from the Rights database and optionally updates related records in DsStorage.
+     * <p/>
+     * This method performs the deletion of a restricted ID by its internal ID in the database
+     * and logs the deletion in the audit log. If specified, it also touches related storage records
      *
-     * @param id value of the id
-     * @param idType
-     * @param platform the platform
-     * @param touchDsStorageRecord
-     * @throws Exception
+     * @param internalId in the database of the restricted ID to be deleted.
+     * @param user the user performing the deletion action, used for audit logging.
+     * @param touchDsStorageRecord a boolean indicating whether to update related storage records.
      */
-    public static void deleteRestrictedId(String id, String idType, String platform, String user, boolean touchDsStorageRecord) throws Exception {
-        BaseModuleStorage.performStorageAction("delete restricted ID",RightsModuleStorage.class, storage -> {
-            ((RightsModuleStorage) storage).deleteRestrictedId(id, idType, platform);
+    public static int deleteRestrictedId(String internalId, String user, boolean touchDsStorageRecord) throws Exception {
+        return BaseModuleStorage.performStorageAction("delete restricted ID",RightsModuleStorage.class, storage -> {
+            // Retrieve object from database
+            RestrictedIdOutputDto idToDelete = ((RightsModuleStorage) storage).getRestrictedIdByInternalId(internalId);
+
+            // Delete entry from database
+            int deletedCount = ((RightsModuleStorage) storage).deleteRestrictedIdByInternalId(internalId);
             if (touchDsStorageRecord) {
-                touchRelatedStorageRecords(id, idType);
+                touchRelatedStorageRecords(idToDelete.getIdValue(), idToDelete.getIdType());
             }
 
-            ChangeDifferenceText change = RightsChangelogGenerator.deleteRestrictedIdChanges(id,idType,platform);
-            AuditLog logEntry = new AuditLog(System.currentTimeMillis(), user,"Delete restricted ID (klausulering)", id, change.getBefore(),"");
+            ChangeDifferenceText change = RightsChangelogGenerator.deleteRestrictedIdChanges(idToDelete.getIdValue(), idToDelete.getIdType(), idToDelete.getPlatform());
+            AuditLog logEntry = new AuditLog(System.currentTimeMillis(), user,"Delete restricted ID (klausulering)", idToDelete.getIdValue(), change.getBefore(),"");
             storage.persistAuditLog(logEntry);
-            log.info("Deleted restriction {} {} {}",id,idType,platform);
-            return null;
+            log.info("Deleted restriction for internal ID: '{}' with idValue: '{}' with idType: '{}' on platform: '{}'.",
+                    internalId, idToDelete.getIdValue(), idToDelete.getIdType(), idToDelete.getPlatform());
+            return deletedCount;
         });
-        log.info("Deleted restricted id:{} idType:{} platform:{}", id, idType, platform);
     }
 
     /**
@@ -218,25 +223,28 @@ public class RightsModuleFacade {
      * @param restrictedIds        list of restricted Ids to be deleted.
      * @param touchDsStorageRecord
      */
-    public static void deleteRestrictedIds(List<RestrictedIdInputDto> restrictedIds, String user, boolean touchDsStorageRecord) {
-        BaseModuleStorage.performStorageAction("delete restricted ID",RightsModuleStorage.class, storage -> {
-            for(RestrictedIdInputDto id : restrictedIds) {
-                ((RightsModuleStorage) storage).deleteRestrictedId(
-                        id.getIdValue(),
-                        id.getIdType(),
-                        id.getPlatform()
-                );
+    public static int deleteRestrictedIds(List<RestrictedIdInputDto> restrictedIds, String user, boolean touchDsStorageRecord) {
+        int deletedRecords = BaseModuleStorage.performStorageAction("delete restricted ID",RightsModuleStorage.class, storage -> {
+            int totalDeleted = 0;
+            for(RestrictedIdInputDto internalId : restrictedIds) {
+                // Get ID for deletion to extract value and type from internal ID
+                RestrictedIdOutputDto idToDelete = ((RightsModuleStorage) storage).getRestrictedIdByInternalId(internalId.getInternalId());
+
+                // Delete each entry by internal ID
+                int deletedCount = ((RightsModuleStorage) storage).deleteRestrictedIdByInternalId(internalId.getInternalId());
                 if (touchDsStorageRecord) {
-                    touchRelatedStorageRecords(id.getIdValue(), id.getIdType());
+                    touchRelatedStorageRecords(idToDelete.getIdValue(), idToDelete.getIdType());
                 }
-                ChangeDifferenceText change = RightsChangelogGenerator.deleteRestrictedIdChanges(id.getIdValue(),id.getIdType(),id.getPlatform());
-                AuditLog logEntry = new AuditLog(System.currentTimeMillis(), user,"Delete restricted ID (klausulering)", id.getIdValue(), change.getBefore(),"");
+                ChangeDifferenceText change = RightsChangelogGenerator.deleteRestrictedIdChanges(idToDelete.getIdValue(), idToDelete.getIdType(), idToDelete.getPlatform());
+                AuditLog logEntry = new AuditLog(System.currentTimeMillis(), user,"Delete restricted ID (klausulering)", idToDelete.getIdValue(), change.getBefore(),"");
                 storage.persistAuditLog(logEntry);
+                totalDeleted += deletedCount;
             }
-            return null;
+            return totalDeleted;
         });
         log.info("Deleted restricted IDs: [{}] ",
                 restrictedIds.stream().map(RestrictedIdInputDto::toString).collect(Collectors.joining(", ")));
+        return deletedRecords;
     }