+++ src/main/java/dk/kb/license/storage/RightsModuleStorage.java
@@ -40,6 +40,8 @@ public class RightsModuleStorage extends BaseModuleStorage{
             RESTRICTED_ID_COMMENT +","+ RESTRICTED_ID_MODIFIED_BY +","+ RESTRICTED_ID_MODIFIED_TIME +")" +
             " VALUES (?,?,?,?,?,?,?)";
     private final String readRestrictedIdQuery = "SELECT * FROM " + RESTRICTED_ID_TABLE + " WHERE " + RESTRICTED_ID_IDVALUE + " = ? AND " + RESTRICTED_ID_IDTYPE + " = ? AND " + RESTRICTED_ID_PLATFORM + " = ? ";
+    private final String readRestrictedIdQueryByInternalId = "SELECT * FROM " + RESTRICTED_ID_TABLE + " WHERE " + RESTRICTED_ID_ID + " = ?";
+
     private final String updateRestrictedIdQuery = "UPDATE " + RESTRICTED_ID_TABLE +" SET "+
             RESTRICTED_ID_PLATFORM +" = ? , " +
             RESTRICTED_ID_COMMENT +" = ? , " +
@@ -49,6 +51,8 @@ public class RightsModuleStorage extends BaseModuleStorage{
             RESTRICTED_ID_IDVALUE +" = ? AND " +
             RESTRICTED_ID_IDTYPE +" = ?" ;
     private final String deleteRestrictedIdQuery = "DELETE FROM " + RESTRICTED_ID_TABLE + " WHERE " + RESTRICTED_ID_IDVALUE + " = ? AND " + RESTRICTED_ID_IDTYPE + " = ? AND " + RESTRICTED_ID_PLATFORM + " = ? ";
+    private final String deleteRestrictedIdByInternalIdQuery = "DELETE FROM " + RESTRICTED_ID_TABLE + " WHERE " + RESTRICTED_ID_ID + " = ?";
+
     private final String allRestrictedIdsQuery = "SELECT * FROM " + RESTRICTED_ID_TABLE + " WHERE " + RESTRICTED_ID_IDTYPE + " LIKE ? AND " + RESTRICTED_ID_PLATFORM + " LIKE ?";
 
     private final String DR_HOLDBACK_RULES_TABLE = "DR_HOLDBACK_RULES";
@@ -149,15 +153,7 @@ public class RightsModuleStorage extends BaseModuleStorage{
             stmt.setString(3, platform);
             ResultSet res = stmt.executeQuery();
             while (res.next()) {
-                RestrictedIdOutputDto output = new RestrictedIdOutputDto();
-                output.setIdValue(res.getString(RESTRICTED_ID_IDVALUE));
-                output.setIdType(res.getString(RESTRICTED_ID_IDTYPE));
-                output.setPlatform(res.getString(RESTRICTED_ID_PLATFORM));
-                output.setComment(res.getString(RESTRICTED_ID_COMMENT));
-                output.setModifiedBy(res.getString(RESTRICTED_ID_MODIFIED_BY));
-                output.setModifiedTime(res.getLong(RESTRICTED_ID_MODIFIED_TIME));
-                output.setModifiedTimeHuman(convertToHumanReadable(output.getModifiedTime()));
-                return output;
+                return createRestrictedIdOutputDtoFromResultSet(res);
             }
             return null;
         } catch (SQLException e) {
@@ -194,7 +190,7 @@ public class RightsModuleStorage extends BaseModuleStorage{
     }
 
     /**
-     * Delete an entry from the restricted IDs table. Also updates the mTime of the related record in ds-storage.
+     * Delete an entry from the restricted IDs table.
      *
      * @param id_value value of the ID
      * @param id_type type of ID
@@ -209,6 +205,44 @@ public class RightsModuleStorage extends BaseModuleStorage{
             stmt.execute();
         } catch (SQLException e) {
             log.error("SQL Exception in delete restricted Id:" + e.getMessage());
+            throw e;
+        }
+    }
+
+
+    /**
+     * Get an entry from the restricted IDs table by internal ID.
+     *
+     * @param internalId to get entry for.
+     * @return a {@link RestrictedIdOutputDto}
+     */
+    public RestrictedIdOutputDto getRestrictedIdByInternalId(String internalId) throws SQLException {
+        try (PreparedStatement stmt = connection.prepareStatement(readRestrictedIdQueryByInternalId)) {
+            stmt.setString(1, internalId);
+            ResultSet res = stmt.executeQuery();
+            while (res.next()) {
+                return createRestrictedIdOutputDtoFromResultSet(res);
+            }
+            return null;
+        } catch (SQLException e) {
+            log.error("SQL Exception in readClause:" + e.getMessage());
+            throw e;
+        }
+    }
+
+    /**
+     * Delete an entry from the restricted IDs table by internal ID.
+     *
+     * @param internalId to delete entry for.
+     */
+    public int deleteRestrictedIdByInternalId(String internalId) throws SQLException {
+        try (PreparedStatement statement = connection.prepareStatement(deleteRestrictedIdByInternalIdQuery)) {
+            statement.setString(1, internalId);
+            int result = statement.executeUpdate();
+            log.info("Deleted '{}' documents by internal ID: '{}'", result, internalId);
+            return result;
+        } catch (SQLException e){
+            log.error("SQL Exception in delete internal ID: " + e.getMessage());
             throw e;
         }
     }
@@ -228,14 +262,7 @@ public class RightsModuleStorage extends BaseModuleStorage{
             ResultSet res = stmt.executeQuery();
             List<RestrictedIdOutputDto> output = new ArrayList<>();
             while (res.next()) {
-                RestrictedIdOutputDto restrictedIdOutputDto = new RestrictedIdOutputDto();
-                restrictedIdOutputDto.setIdValue(res.getString(RESTRICTED_ID_IDVALUE));
-                restrictedIdOutputDto.setIdType(res.getString(RESTRICTED_ID_IDTYPE));
-                restrictedIdOutputDto.setPlatform(res.getString(RESTRICTED_ID_PLATFORM));
-                restrictedIdOutputDto.setComment(res.getString(RESTRICTED_ID_COMMENT));
-                restrictedIdOutputDto.setModifiedBy(res.getString(RESTRICTED_ID_MODIFIED_BY));
-                restrictedIdOutputDto.setModifiedTime(res.getLong(RESTRICTED_ID_MODIFIED_TIME));
-                restrictedIdOutputDto.setModifiedTimeHuman(convertToHumanReadable(restrictedIdOutputDto.getModifiedTime()));
+                RestrictedIdOutputDto restrictedIdOutputDto = createRestrictedIdOutputDtoFromResultSet(res);
                 output.add(restrictedIdOutputDto);
             }
             return output;
@@ -525,6 +552,23 @@ public class RightsModuleStorage extends BaseModuleStorage{
         }
         connection.commit();
         log.info("All tables cleared for unittest");
+    }
+
+    /**
+     * Create a {@link RestrictedIdOutputDto} from a ResultSet, which should contain all needed values for the DTO
+     * @param resultSet containing values from the backing Rights database
+     */
+    private RestrictedIdOutputDto createRestrictedIdOutputDtoFromResultSet(ResultSet resultSet) throws SQLException {
+        RestrictedIdOutputDto output = new RestrictedIdOutputDto();
+        output.setInternalId(resultSet.getString(RESTRICTED_ID_ID));
+        output.setIdValue(resultSet.getString(RESTRICTED_ID_IDVALUE));
+        output.setIdType(resultSet.getString(RESTRICTED_ID_IDTYPE));
+        output.setPlatform(resultSet.getString(RESTRICTED_ID_PLATFORM));
+        output.setComment(resultSet.getString(RESTRICTED_ID_COMMENT));
+        output.setModifiedBy(resultSet.getString(RESTRICTED_ID_MODIFIED_BY));
+        output.setModifiedTime(resultSet.getLong(RESTRICTED_ID_MODIFIED_TIME));
+        output.setModifiedTimeHuman(convertToHumanReadable(output.getModifiedTime()));
+        return output;
     }
 
     /*